requirements.txt
# LangGraph Stock Analysis - Python Dependencies
# Graduation Thesis Project - February 2026

# ============================================================================
# LangGraph & LLM
# ============================================================================
langgraph>=0.0.55
langchain>=0.1.20
groq>=0.4.0

# ============================================================================
# Data & APIs
# ============================================================================
polygon-api-client>=1.12.0
finnhub-python>=2.4.19
yfinance>=0.2.28
pandas>=2.2.0
numpy>=1.26.0

# ============================================================================
# Technical Analysis
# ============================================================================
pandas-ta>=0.4.71b0

# ============================================================================
# Machine Learning / NLP
# ============================================================================
transformers>=4.33.0
torch>=2.0.0
scikit-learn>=1.4.0

# ============================================================================
# Visualization
# ============================================================================
plotly>=5.17.0
streamlit>=1.28.0

# ============================================================================
# Statistics
# ============================================================================
scipy>=1.11.0

# ============================================================================
# Utilities
# ============================================================================
python-dotenv>=1.0.0
aiohttp>=3.9.0

# ============================================================================
# Testing
# ============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# ============================================================================
# Installation Instructions
# ============================================================================
# 1. Create virtual environment: python -m venv venv
# 2. Activate: source venv/bin/activate (Mac/Linux) or venv\Scripts\activate (Windows)
# 3. Install: pip install -r requirements.txt
